Add 'metrics' callback to Kinesis consumer interface
This adds an optional metrics callback.

The only metric sent is MillisBehindLatest which can be tracked to gauge if the
processor is falling behind and the app needs to be scaled.

Doing it this way makes this a non-breaking change and allows additional
metrics to be sent to the callback in the future without breaking it.  For
example, another metric might be how long the GetRecords call took.